The main goal was to bring the last correction fixes related to the type change of code_num from int to bigint
We forgot to bring modification also in the collections table on the code_last_value field and in the trigger updating this field

fix #1794

Conflicts:
	data/db/changes/0068-codes_type_correction_for_imports.sql

+++ data/db/createfunctions.sql
@@ -3772,7 +3772,7 @@ CREATE OR REPLACE FUNCTION check_auto_increment_code_in_spec() RETURNS trigger
 AS $$
 DECLARE 
   col collections%ROWTYPE;
-  number integer ;
+  number BIGINT ;
 BEGIN
   IF TG_OP != 'DELETE' THEN
     IF NEW.referenced_relation = 'specimens' THEN
@@ -3780,7 +3780,7 @@ BEGIN
       IF FOUND THEN
         IF NEW.code_category = 'main' THEN
           IF isnumeric(NEW.code) THEN 
-            number := NEW.code::integer ;
+            number := NEW.code::bigint ;
             IF number > col.code_last_value THEN
               UPDATE collections set code_last_value = number WHERE id=col.id ;
             END IF;
@@ -3813,7 +3813,7 @@ BEGIN
         ELSEIF TG_OP = 'UPDATE' THEN
           IF OLD.code_category = 'main' THEN
             IF isnumeric(OLD.code) THEN 
-              number := OLD.code::integer ;
+              number := OLD.code::bigint ;
               IF number = col.code_last_value THEN
                 UPDATE collections 
                 SET code_last_value = (SELECT max(code_num)
